+++ views/auth/login.handlebars
@@ -1,19 +1,17 @@
-<div class="container">
-  <h1 class="my-5 text-left">Login</h1>
-  <div class="row">
-    <div class="col-lg-12">
-      <div class="row text-center my-5">
-        <div class="col-lg-6 ">
-          <h2>For Donors</h2>
-          <p class="text-muted">Sign In and start donating to millions of users. </p>
-          <a href="/auth" class="btn btn-outline-secondary" id='donor_login'>Login</a>
-        </div>
-        <div class="col-lg-6">
-          <h2>For Recipient</h2>
-          <p class="text-muted">Sign In and receive good from millions of Donors </p>
-          <a href="/auth" class="btn btn-outline-success" id='rep_login'>Login</a>
-        </div>
-      </div>
+<div class="container col-lg-4 my-5">
+  <h2 class="page_title">Login</h2>
+  <form id="login-form" method="post" action="/auth/login">
+    <div class="form-group">
+      <label for="email">Email:</label>
+      <input type="email" class="form-control" id="email" placeholder="Enter email" name="email" autofocus required>
     </div>
-  </div>
+    <div class="form-group">
+      <label for="pwd">Password:</label>
+      <input type="password" class="form-control" id="pwd" placeholder="Enter password" name="password" required>
+    </div>
+    <div class="form-group">
+      <a href="/auth/signup">Sign up?</a>
+    </div>
+    <button type="submit" class="btn btn-primary">Submit</button>
+  </form>
 </div>
