+++ views/auth/signup.handlebars
@@ -56,6 +56,6 @@
       </div>
     </div> --}}
     <button type="submit" class="btn btn-primary">Sign Up</button>
-    <a href="/users/login">Sign in?</a>
+    <a href="/auth/login">Sign in?</a>
   </form>
 </div>
