+++ views/layouts/main.handlebars
@@ -26,9 +26,9 @@
       <ul class="navbar-nav">
         <li class="nav-item">
           {{#if authenticated}}
-            <a href="/users/logout" class="btn btn-outline-success mr-sm-2">Logout</a>
+            <a href="/auth/logout" class="btn btn-outline-success mr-sm-2">Logout</a>
           {{else}}
-            <a href="/users/login" class="btn btn-outline-success mr-sm-2">Login</a>
+            <a href="/auth/login" class="btn btn-outline-success mr-sm-2">Login</a>
           {{/if}}
         </li>
         <li class="nav-item">
