+++ views/static/home.handlebars
@@ -9,14 +9,14 @@
             <p class="text-muted">Donation can be made easy by simply uploading an item to the page. Eveyone can donate
               or
               become a donor. </p>
-            <a href="/users/new" class="btn btn-outline-secondary" id="donor_login">Start Donating</a>
+            <a href="/auth/signup" class="btn btn-outline-secondary">Start Donating</a>
           </div>
           <div class="col-12 col-md-6 text-center">
             <h2>For Recipient</h2>
             <p class="text-muted">This platform is built for anyone who is in need. Are you in need? Time to shop for $0
             </p>
             <br>
-            <a href="/users/new" class="btn btn-outline-success" id='rep_login'>Sign Up & Receive</a>
+            <a href="/auth/signup" class="btn btn-outline-success">Sign Up & Receive</a>
           </div>
         </div>
       </div>
