--- views/users/login.handlebars
@@ -1,17 +1,0 @@
-<div class="container col-lg-4 my-5">
-  <h2 class="page_title">Login</h2>
-  <form id="login-form" method="post" action="/users/login">
-    <div class="form-group">
-      <label for="email">Email:</label>
-      <input type="email" class="form-control" id="email" placeholder="Enter email" name="email" autofocus required>
-    </div>
-    <div class="form-group">
-      <label for="pwd">Password:</label>
-      <input type="password" class="form-control" id="pwd" placeholder="Enter password" name="password" required>
-    </div>
-    <div class="form-group">
-      <a href="/users/new">Sign up?</a>
-    </div>
-    <button type="submit" class="btn btn-primary">Submit</button>
-  </form>
-</div>
